requirements.txt
#------------------------------------------------------------------------------
#--- !!!   IMPORTANT   !!!
#    This is a Collection of all Project Dependencies.
#    !!! DO NOT REMOVE !!!
#------------------------------------------------------------------------------
#--- Upgrade all
#    pip freeze --local | grep -v "^\-e" | cut -d = -f 1 | xargs -n1 pip install -U

#------------------------------------------------------------------------------
# "#"
APScheduler==3.5.1
BeautifulSoup==3.2.1
Django<2.0
Fabric==2.0.1
Markdown==2.6.11
Pillow==5.1.0

#------------------------------------------------------------------------------
# "A"

#------------------------------------------------------------------------------
# "B"
boto==2.48.0

#------------------------------------------------------------------------------
# "C"
coverage==4.5.1

#------------------------------------------------------------------------------
# "D"
djangorestframework==3.8.2
djangorestframework-jsonp==1.0.2

#------------------------------------------------------------------------------
# "django-*"
django-admin-rangefilter==0.3.6
django-admin-sortable2==0.6.19
django-annoying==0.10.4
django-auth-remember==0.3
django-autoslug==1.9.3
django-bootstrap3-datetimepicker==2.2.3
django-bower==5.2.0
django-cache-machine==1.0.0
django-ckeditor==5.4.0
django-compressor==2.2
django-countries==5.3
django-cyborg==0.2.1
django-debug-toolbar==1.9.1
django-easy-pdf==0.1.1
django-easy-timezones==0.8.0
django-extensions==2.0.7
django-filter==1.1.0
django-formset-js==0.5.0
django-grappelli==2.11.1

git+git://github.com/django-haystack/django-haystack.git@master

django-imagekit==4.0.2
django-ipware==2.1.0
django-mptt==0.9.0
django-mptt-admin==0.5.1
django-papertrail==1.1.8
django-passwords==0.3.12
django-phonenumber-field==2.0.0
django-rest-swagger==2.2.0
django-rosetta==0.8.1
django-secure==1.0.1

git+git://github.com/asuvorov/django-select-multiple-field.git@master

django-seo2==1.0.1
django-simple-captcha==0.5.7
django-sslserver==0.20
django-storages==1.6.6
django-taggit==0.22.2
django-taggit-templatetags2==1.6.1
django-timezone-field==2.1
django-twitter-tag==1.2.1
django-url-tools==0.0.8
django-wysiwyg==0.8.0

#------------------------------------------------------------------------------
# "E"
elasticsearch==6.2.0

#------------------------------------------------------------------------------
# "F"

#------------------------------------------------------------------------------
# "G"

#------------------------------------------------------------------------------
# "H"

#------------------------------------------------------------------------------
# "I"
ipaddress==1.0.22
ipython==5.7.0

#------------------------------------------------------------------------------
# "J"
jsonfield==2.0.2

#------------------------------------------------------------------------------
# "K"

#------------------------------------------------------------------------------
# "L"
lxml==4.2.1

#------------------------------------------------------------------------------
# "M"
mock==2.0.0
mysql-python==1.2.5

#------------------------------------------------------------------------------
# "N"

#------------------------------------------------------------------------------
# "O"

#------------------------------------------------------------------------------
# "P"
pendulum==2.0.1
pep8==1.7.1
pep257==0.7.0
progress==1.3
pyflakes==2.0.0
pyjwkest==1.4.0
pylint==1.9.1
pylint-django==0.11
pytz==2018.4

#------------------------------------------------------------------------------
# "python-*"
python-magic==0.4.15
python-memcached==1.59
python-social-auth==0.3.6

#------------------------------------------------------------------------------
# "Q"

#------------------------------------------------------------------------------
# "R"
requests==2.18.4

#------------------------------------------------------------------------------
# "S"
selenium==3.12.0
sendgrid-django==4.2.0
social-auth-app-django==2.1.0
supervisor==3.3.4

#------------------------------------------------------------------------------
# "T"
termcolor==1.1.0
tzlocal==1.4

#------------------------------------------------------------------------------
# "U"
urllib3==1.22
uWSGI==2.0.17

#------------------------------------------------------------------------------
# "V"

#------------------------------------------------------------------------------
# "W"
whitenoise==3.3.1

#------------------------------------------------------------------------------
# "X"
xlwt==1.3.0

#------------------------------------------------------------------------------
# "Y"

#------------------------------------------------------------------------------
# "Z"
